refactor|chore(home-14): add of react-markdown
