+++ package.json
@@ -37,6 +37,7 @@
         "@mui/material": "7.3.4",
         "react": "19.2.0",
         "react-dom": "19.2.0",
+        "react-markdown": "^10.1.0",
         "react-router-dom": "7.9.3"
     },
     "devDependencies": {
